Client authentication, menus, redux state, localstorage
Quite a few changes here.

I completed (mostly I think) of the authentication client side, also what
is shown based on if we authenticated or not.
Added auth reducer/redux logic to keep state
Added auth actions to execute b4 the state change.
Added localstorage to store the expiration time so we can logout
automaticaly once the expiration is hit.
Added routing and redirection depending on if we are authenticated.
Added place holders for Call for Proposal and Admin Proposals pages.
Added to the menu tool bar links to the above pages
Added logic so if we reauthenticate, say after a reload we are
redirected back to the page we were.