Django authorization REST endpoints
I've added a bunch of ways to do authentication. Finally I decided to
use knox authentication. A package that already handles a bunch of
things for your. For instance expiry date, expiry date format, number of
tokens issued etc.
Also the knox project is available in git and the test code is a good
source of information.