Creating the CfP end point
Call For Proposals API end point sets the open and close times for the
given semester.
There are four end points.
A CfpViewSet that accepts creation, updates and list queries of the CfP
table.
And a CfPCurrentSem API that returns a specific json structure the
react javascript app requires.
Basically
   currenSem:{current cfp semester details},
   semList:[{<cfp semester details}]

On request if the current semester is not yet present it'll get created.
Extensive unit testing is provided.